backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.12.0
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.8.5
PyMySQL==1.0.2
python-dateutil==1.5
python-dotenv==0.14.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.0.30
whitenoise==5.3.0